refactor: 검색 시 모든 게시판의 글 출력
