+++ src/components/Community/Post/Content/index.jsx
@@ -9,7 +9,7 @@ import authIcon from "../../../../assets/icons/common/Community/CheckCircle.svg"
 import PostMenuBar from "../MenuBar/PostMenubar";
 import { useNavigate } from "react-router-dom";
 import useComment from "../../../../hooks/useComment";
-import {scrapPost, postLike} from "../../../../api/post";
+import {scrapPost, postLike} from "../../../../api/postAPI";
 import bookmarkFillIcon from "../../../../assets/icons/common/Community/Scrap.svg";
 import { formatDate } from "../../../../utils/dateUtils";
